Feat: Standardize dropdown fields and add a new image field
This commit improves the appearance and functionality of dropdown menus in the CreateCar form. Similar to the Drive Type example, consistent styling and icon integration have been achieved for all enum and brand selection fields.

Furthermore, a `BigImageUrl` field has been added to the DTO to allow users to add both a cover image and additional detail images for the vehicle, and the UI components for this field have been integrated into the CreateCar.cshtml page.

**Changes:**
- A more user-friendly interface with icons has been created for all select (dropdown) fields using the `Select2` library.
- The `BigImageUrl` property has been added to `CreateCarDto`.
- A new input field for `BigImageUrl`, styled identically to `CoverImageUrl`, has been added to the CreateCar.cshtml page.
- CSS and JavaScript have been revised to reflect the new field and updated dropdown styles.
- Sample data has been added for the **Ford Fiesta** (for Model, Year, Trunk Volume, etc.).
- Possible causes of the `400 Bad Request` error received when adding a vehicle (specifically, newly added mandatory fields such as `BigImageUrl` not yet being met on the API side) will be investigated.